# source-code/cuongpt-next-tutor/src/main/java/com/cuongpt/nexttutor/api/DefaultApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.cuongpt.nexttutor.api;

import com.cuongpt.nexttutor.model.dto.request.GetHintRequestDTO;
import com.cuongpt.nexttutor.model.dto.response.GetHintResponseDTO;
import com.cuongpt.nexttutor.model.dto.request.InteractRequestDTO;
import com.cuongpt.nexttutor.model.dto.response.InteractResponseDTO;
import com.cuongpt.nexttutor.model.dto.request.LoginRequestDTO;
import com.cuongpt.nexttutor.model.dto.response.LoginResponseDTO;
import com.cuongpt.nexttutor.model.dto.response.PronunciationAssessmentResponseDTO;
import com.cuongpt.nexttutor.model.dto.QuizDTO;
import com.cuongpt.nexttutor.model.dto.request.StartSessionRequestDTO;
import com.cuongpt.nexttutor.model.dto.response.StartSessionResponseDTO;
import com.cuongpt.nexttutor.model.dto.response.SttResponseDTO;
import com.cuongpt.nexttutor.model.dto.request.SubmitQuizRequestDTO;
import com.cuongpt.nexttutor.model.dto.response.SubmitQuizResponseDTO;
import com.cuongpt.nexttutor.model.dto.request.TtsRequestDTO;
import com.cuongpt.nexttutor.model.dto.TutorDTO;
import java.util.UUID;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;

import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-03T18:26:26.686856118+07:00[Asia/Ho_Chi_Minh]", comments = "Generator version: 7.12.0")
@Validated
@Tag(name = "Default", description = "the Default API")
public interface DefaultApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/v1/ai/assessment/pronunciation : Pronunciation Assessment
     *
     * @param audio  (optional)
     * @param targetText  (optional)
     * @return Assessment result (status code 200)
     */
    @Operation(
        operationId = "apiV1AiAssessmentPronunciationPost",
        summary = "Pronunciation Assessment",
        responses = {
            @ApiResponse(responseCode = "200", description = "Assessment result", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PronunciationAssessmentResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/ai/assessment/pronunciation",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    default ResponseEntity<PronunciationAssessmentResponseDTO> apiV1AiAssessmentPronunciationPost(
        @Parameter(name = "audio", description = "") @RequestPart(value = "audio", required = false) MultipartFile audio,
        @Parameter(name = "targetText", description = "") @Valid @RequestParam(value = "targetText", required = false) String targetText
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"score\" : 0, \"phonemes\" : [ { \"isGood\" : true, \"phoneme\" : \"phoneme\", \"suggestion\" : \"suggestion\" }, { \"isGood\" : true, \"phoneme\" : \"phoneme\", \"suggestion\" : \"suggestion\" } ], \"overallFeedback\" : \"overallFeedback\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/ai/stt : Speech to Text
     *
     * @param file  (optional)
     * @return Transcribed text (status code 200)
     */
    @Operation(
        operationId = "apiV1AiSttPost",
        summary = "Speech to Text",
        responses = {
            @ApiResponse(responseCode = "200", description = "Transcribed text", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SttResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/ai/stt",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    default ResponseEntity<SttResponseDTO> apiV1AiSttPost(
        @Parameter(name = "file", description = "") @RequestPart(value = "file", required = false) MultipartFile file
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"confidence\" : 0.8008281904610115, \"text\" : \"text\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/ai/tts : Text to Speech
     *
     * @param ttsRequestDTO  (required)
     * @return Audio stream (status code 200)
     */
    @Operation(
        operationId = "apiV1AiTtsPost",
        summary = "Text to Speech",
        responses = {
            @ApiResponse(responseCode = "200", description = "Audio stream", content = {
                @Content(mediaType = "audio/mpeg", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/ai/tts",
        produces = { "audio/mpeg" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<org.springframework.core.io.Resource> apiV1AiTtsPost(
        @Parameter(name = "TtsRequestDTO", description = "", required = true) @Valid @RequestBody TtsRequestDTO ttsRequestDTO
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/auth/login : Login with OAuth2 provider
     *
     * @param loginRequestDTO  (required)
     * @return Successful login (status code 200)
     */
    @Operation(
        operationId = "apiV1AuthLoginPost",
        summary = "Login with OAuth2 provider",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful login", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoginResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<LoginResponseDTO> apiV1AuthLoginPost(
        @Parameter(name = "LoginRequestDTO", description = "", required = true) @Valid @RequestBody LoginRequestDTO loginRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"user\" : { \"role\" : \"LEARNER\", \"avatarUrl\" : \"avatarUrl\", \"displayName\" : \"displayName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"token\" : \"token\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/quizzes : Get a quiz based on criteria
     *
     * @param level  (optional)
     * @param skill  (optional)
     * @param topic  (optional)
     * @return Quiz generated (status code 200)
     */
    @Operation(
        operationId = "apiV1QuizzesGet",
        summary = "Get a quiz based on criteria",
        responses = {
            @ApiResponse(responseCode = "200", description = "Quiz generated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = QuizDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/quizzes",
        produces = { "application/json" }
    )
    
    default ResponseEntity<QuizDTO> apiV1QuizzesGet(
        @Parameter(name = "level", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "level", required = false) String level,
        @Parameter(name = "skill", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "skill", required = false) String skill,
        @Parameter(name = "topic", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "topic", required = false) String topic
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"quizId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"questions\" : [ { \"audioUrl\" : \"audioUrl\", \"options\" : [ \"options\", \"options\" ], \"id\" : \"id\", \"type\" : \"MULTIPLE_CHOICE\", \"content\" : \"content\" }, { \"audioUrl\" : \"audioUrl\", \"options\" : [ \"options\", \"options\" ], \"id\" : \"id\", \"type\" : \"MULTIPLE_CHOICE\", \"content\" : \"content\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/quizzes/{quizId}/submit : Submit quiz answers
     *
     * @param quizId  (required)
     * @param submitQuizRequestDTO  (required)
     * @return Quiz results (status code 200)
     */
    @Operation(
        operationId = "apiV1QuizzesQuizIdSubmitPost",
        summary = "Submit quiz answers",
        responses = {
            @ApiResponse(responseCode = "200", description = "Quiz results", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SubmitQuizResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/quizzes/{quizId}/submit",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<SubmitQuizResponseDTO> apiV1QuizzesQuizIdSubmitPost(
        @Parameter(name = "quizId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("quizId") UUID quizId,
        @Parameter(name = "SubmitQuizRequestDTO", description = "", required = true) @Valid @RequestBody SubmitQuizRequestDTO submitQuizRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"score\" : 0, \"totalQuestions\" : 6, \"details\" : [ { \"questionId\" : \"questionId\", \"explanation\" : \"explanation\", \"isCorrect\" : true }, { \"questionId\" : \"questionId\", \"explanation\" : \"explanation\", \"isCorrect\" : true } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sessions/{sessionId}/hint : Get a hint for the current context
     *
     * @param sessionId  (required)
     * @param getHintRequestDTO  (optional)
     * @return Hint provided (status code 200)
     */
    @Operation(
        operationId = "apiV1SessionsSessionIdHintPost",
        summary = "Get a hint for the current context",
        responses = {
            @ApiResponse(responseCode = "200", description = "Hint provided", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetHintResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sessions/{sessionId}/hint",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<GetHintResponseDTO> apiV1SessionsSessionIdHintPost(
        @Parameter(name = "sessionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("sessionId") UUID sessionId,
        @Parameter(name = "GetHintRequestDTO", description = "") @Valid @RequestBody(required = false) GetHintRequestDTO getHintRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"hintText\" : \"hintText\", \"hintAudioUrl\" : \"hintAudioUrl\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sessions/{sessionId}/interact : Interact with AI Tutor
     *
     * @param sessionId  (required)
     * @param interactRequestDTO  (required)
     * @return AI Response (status code 200)
     */
    @Operation(
        operationId = "apiV1SessionsSessionIdInteractPost",
        summary = "Interact with AI Tutor",
        responses = {
            @ApiResponse(responseCode = "200", description = "AI Response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InteractResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sessions/{sessionId}/interact",
        produces = { "application/json" },
        consumes = { "application/json", "multipart/form-data" }
    )
    
    default ResponseEntity<InteractResponseDTO> apiV1SessionsSessionIdInteractPost(
        @Parameter(name = "sessionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("sessionId") UUID sessionId,
        @Parameter(name = "InteractRequestDTO", description = "", required = true) @Valid @RequestBody InteractRequestDTO interactRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"feedback\" : { \"score\" : 0, \"correction\" : \"correction\", \"pronunciationErrors\" : [ \"{}\", \"{}\" ] }, \"aiResponse\" : { \"audioUrl\" : \"audioUrl\", \"text\" : \"text\" } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sessions/start : Start a new learning session
     *
     * @param startSessionRequestDTO  (required)
     * @return Session started (status code 200)
     */
    @Operation(
        operationId = "apiV1SessionsStartPost",
        summary = "Start a new learning session",
        responses = {
            @ApiResponse(responseCode = "200", description = "Session started", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = StartSessionResponseDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sessions/start",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<StartSessionResponseDTO> apiV1SessionsStartPost(
        @Parameter(name = "StartSessionRequestDTO", description = "", required = true) @Valid @RequestBody StartSessionRequestDTO startSessionRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"initialMessage\" : { \"audioUrl\" : \"audioUrl\", \"text\" : \"text\" }, \"sessionId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/tutors : Get available AI tutors
     *
     * @return List of tutors (status code 200)
     */
    @Operation(
        operationId = "apiV1TutorsGet",
        summary = "Get available AI tutors",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of tutors", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TutorDTO.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/tutors",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<TutorDTO>> apiV1TutorsGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"avatarUrl\" : \"avatarUrl\", \"personalityPrompt\" : \"personalityPrompt\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\" }, { \"avatarUrl\" : \"avatarUrl\", \"personalityPrompt\" : \"personalityPrompt\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}